+++ cadwyn/changelogs.py
@@ -12,9 +12,12 @@ from fastapi._compat import (
     get_definitions,
     get_schema_from_model_field,
 )
+from fastapi.applications import get_openapi
 from fastapi.openapi.constants import REF_TEMPLATE
-from fastapi.openapi.utils import get_fields_from_routes
+from fastapi.openapi.utils import get_fields_from_routes, get_openapi_operation_parameters
+from fastapi.params import Param
 from pydantic import BaseModel, Field, RootModel
+from pydantic.fields import FieldInfo
 from typing_extensions import assert_never
 
 from cadwyn import (
@@ -67,41 +70,67 @@ def _convert_version_change_instruction_to_changelog_entry(
                 methods=cast(Any, instruction.endpoint_methods),
             )
         case EndpointHadInstruction():
-            ...
+            path: str
+            response_model: Any
+            status_code: int
+            tags: list[str | Enum]
+            summary: str
+            description: str
+            response_description: str
+            responses: dict[int | str, dict[str, Any]]
+            deprecated: bool
+            methods: list[str]
+            operation_id: str
+            include_in_schema: bool
+            name: str
+
         case FieldHadInstruction():
+            old_field_name = _get_older_field_name(instruction.schema, instruction.name, generator_from_older_version)
 
-            class CadwynModifiedFieldAttribute(BaseModel):
-                name: str
-                old_value: Any
-                new_value: Any
-
-            model = generator_from_newer_version._get_wrapper_for_model(instruction.schema)
-
-            modified_fields: list[CadwynModifiedFieldAttribute] = []
             if instruction.new_name is not Sentinel:
-                modified_fields.append(
-                    CadwynModifiedFieldAttribute(
-                        name="name", old_value=instruction.new_name, new_value=instruction.name
+                attribute_changes = [
+                    CadwynAttributeChange(
+                        name="name",
+                        status=CadwynAttributeChangeStatus.changed,
+                        old_value=old_field_name,
+                        new_value=instruction.name,
                     )
+                ]
+            else:
+                attribute_changes = []
+
+            newer_model = generator_from_newer_version[instruction.schema]
+            older_model = generator_from_older_version[instruction.schema]
+
+            newer_field_openapi = _get_openapi_representation_of_a_field(newer_model, instruction.name)
+            older_field_openapi = _get_openapi_representation_of_a_field(older_model, old_field_name)
+
+            attribute_changes += [
+                CadwynAttributeChange(
+                    name=key,
+                    status=CadwynAttributeChangeStatus.changed
+                    if key in newer_field_openapi
+                    else CadwynAttributeChangeStatus.removed,
+                    old_value=old_value,
+                    new_value=newer_field_openapi.get(key),
                 )
-            if instruction.type is not Sentinel:
-                modified_fields.append(
-                    CadwynModifiedFieldAttribute(
-                        name="type", old_value=instruction.type, new_value=model.fields[instruction.name].annotation
-                    )
-                )
-            for attr_name in instruction.field_changes.__dataclass_fields__:
-                attr_value = getattr(instruction.field_changes, attr_name)
-                if attr_value is not Sentinel:
-                    modified_fields.append(
-                        CadwynModifiedFieldAttribute(
-                            name=attr_name,
-                            old_value=attr_value,
-                            new_value=model.fields[instruction.name].passed_field_attributes[attr_name],
-                        )
-                    )
+                for key, old_value in older_field_openapi.items()
+                if old_value != newer_field_openapi.get(key)
+            ]
+
+            return CadwynFieldAttributesWereChangedChangelogEntry(
+                models=_get_affected_model_names(instruction, generator_from_newer_version, schemas_from_last_version),
+                field=old_field_name,
+                attribute_changes=attribute_changes,
+            )
+
         case FieldDidntHaveInstruction():
-            ...
+            newer_model = generator_from_newer_version[instruction.schema]
+            older_model = generator_from_older_version[instruction.schema]
+            newer_field_openapi = _get_openapi_representation_of_a_field(newer_model, instruction.name)
+            older_field_openapi = _get_openapi_representation_of_a_field(older_model, old_field_name)
+
+            _get_openapi_representation_of_a_field(model, instruction.name)
         case EnumDidntHaveMembersInstruction():
             enum = generator_from_newer_version._get_wrapper_for_model(instruction.enum)
 
@@ -116,11 +145,13 @@ def _convert_version_change_instruction_to_changelog_entry(
             return CadwynEnumMembersWereChangedChangelogEntry(
                 enum=new_enum.__name__,
                 member_changes=[
-                    CadwynEnumMemberChange(
+                    CadwynAttributeChange(
                         name=name,
                         old_value=old_enum.__members__.get(name),
                         new_value=new_enum.__members__.get(name),
-                        status=CadwynEnumStatus.changed if name in new_enum.__members__ else CadwynEnumStatus.removed,
+                        status=CadwynAttributeChangeStatus.changed
+                        if name in new_enum.__members__
+                        else CadwynAttributeChangeStatus.removed,
                     )
                     for name in instruction.members
                 ],
@@ -145,17 +176,26 @@ def _convert_version_change_instruction_to_changelog_entry(
             return CadwynSchemaFieldWasAddedChangelogEntry(
                 models=affected_model_names,
                 field=instruction.name,
-                field_info=_get_openapi_representation_of_a_field(
-                    ModelField(model.model_fields[instruction.name], instruction.name)
-                ),
+                field_info=_get_openapi_representation_of_a_field(model, instruction.name),
             )
         case staticmethod() | ValidatorDidntExistInstruction() | ValidatorExistedInstruction():
             return []
     assert_never(instruction)
 
 
+def _get_older_field_name(
+    schema: type[BaseModel], new_field_name: str, generator_from_older_version: SchemaGenerator
+) -> str:
+    older_model_wrapper = generator_from_older_version._get_wrapper_for_model(schema)
+    newer_names_mapping = {
+        field.name_from_newer_version: old_name for old_name, field in older_model_wrapper.fields.items()
+    }
+    old_field_name = newer_names_mapping[new_field_name]
+    return old_field_name
+
+
 def _get_affected_model_names(
-    instruction: FieldExistedAsInstruction | FieldDidntExistInstruction,
+    instruction: FieldExistedAsInstruction | FieldDidntExistInstruction | FieldHadInstruction,
     generator_from_newer_version: SchemaGenerator,
     schemas_from_last_version: list[ModelField],
 ):
@@ -200,16 +240,20 @@ def _get_all_pydantic_models_from_generic(annotation: Any) -> list[type[BaseMode
     return models
 
 
-def _get_openapi_representation_of_a_field(field: ModelField):
-    model_name_map = get_compat_model_name_map([field])
+def _get_openapi_representation_of_a_field(model: type[BaseModel], field_name: str) -> dict:
+    class CadwynDummyModelForRepresentation(BaseModel):
+        my_field: model
+
+    model_name_map = get_compat_model_name_map([CadwynDummyModelForRepresentation.model_fields["my_field"]])
     schema_generator = GenerateJsonSchema(ref_template=REF_TEMPLATE)
-    field_mapping, _ = get_definitions(
-        fields=[field],
+    field_mapping, definitions = get_definitions(
+        fields=[ModelField(CadwynDummyModelForRepresentation.model_fields["my_field"], "my_field")],
         schema_generator=schema_generator,
         model_name_map=model_name_map,
         separate_input_output_schemas=False,
     )
-    return list(field_mapping.values())[0]
+
+    return definitions[model.__name__]["properties"][field_name]
 
 
 class ChangelogEntryType(StrEnum):
@@ -219,6 +263,29 @@ class ChangelogEntryType(StrEnum):
     enum_members_removed = "enum.members.removed"
     schema_field_removed = "schema.field.removed"
     schema_field_added = "schema.field.added"
+    schema_field_attributes_changed = "schema.field.attributes.changed"
+    schema_field_attributes_added = "schema.field.attributes.added"
+
+
+class CadwynAttributeChangeStatus(StrEnum):
+    changed = auto()
+    removed = auto()
+
+
+class CadwynAttributeChange(BaseModel):
+    name: str
+    status: CadwynAttributeChangeStatus
+    old_value: Any
+    new_value: Any
+
+
+class CadwynFieldAttributesWereChangedChangelogEntry(BaseModel):
+    type: Literal[ChangelogEntryType.schema_field_attributes_changed] = (
+        ChangelogEntryType.schema_field_attributes_changed
+    )
+    models: list[str]
+    field: str
+    attribute_changes: list[CadwynAttributeChange]
 
 
 class CadwynModelInfo(BaseModel):
@@ -249,18 +316,6 @@ class CadwynEnumMember(BaseModel):
     value: Any
 
 
-class CadwynEnumStatus(StrEnum):
-    changed = auto()
-    removed = auto()
-
-
-class CadwynEnumMemberChange(BaseModel):
-    name: str
-    status: CadwynEnumStatus
-    old_value: Any
-    new_value: Any
-
-
 class CadwynEnumMembersWereAddedChangelogEntry(BaseModel):
     type: Literal[ChangelogEntryType.enum_members_added] = ChangelogEntryType.enum_members_added
     enum: str
@@ -270,7 +325,7 @@ class CadwynEnumMembersWereAddedChangelogEntry(BaseModel):
 class CadwynEnumMembersWereChangedChangelogEntry(BaseModel):
     type: Literal[ChangelogEntryType.enum_members_removed] = ChangelogEntryType.enum_members_removed
     enum: str
-    member_changes: list[CadwynEnumMemberChange]
+    member_changes: list[CadwynAttributeChange]
 
 
 class HTTPMethod(StrEnum):
@@ -322,6 +377,7 @@ CadwynVersionChangeInstruction = RootModel[
     | CadwynEndpointWasRemovedChangelogEntry
     | CadwynSchemaFieldWasRemovedChangelogEntry
     | CadwynSchemaFieldWasAddedChangelogEntry
+    | CadwynFieldAttributesWereChangedChangelogEntry
 ]